improved documentation

+++ launcher/WCE_from_SAD_outputs.sh
@@ -1,3 +1,7 @@
+# Runs WCE for .wav files using the corresponding speech activity detection (SAD)
+# segments defined in .rttm files named as <sadname>_<wavname>.rttm , all located
+# in the same directory <dirname>.
+
 if [ $# -lt 1 ]; then
   echo "Usage: sh WCE_from_SAD_outputs.sh <dirname> <sadname>"
   echo "where dirname is a folder containing audio .wav files and the corresponding .rttm files"
@@ -10,17 +14,20 @@ fi
 SCRIPT_DIR=$(dirname "$0")
 DATA_FOLDER=$1
 
-#if [ -n "$2" ]; then
 if [ "$#" -ne 2 ]; then
-SADNAME="opensmile"
+SADNAME="tocomboSad"
 else
 SADNAME=$2
 fi
 
+if [ -d "${DATA_FOLDER}/wav_tmp/" ]; then
 rm ${DATA_FOLDER}/wav_tmp/*.wav
+fi
 
 python /home/vagrant/repos/WCE_VM/aux_VM/rttm_to_wavs.py ${DATA_FOLDER} ${SADNAME}
 
+echo "Running WCE..."
 sh /home/vagrant/launcher/estimateWCE.sh  ${DATA_FOLDER}/wav_tmp/ ${DATA_FOLDER}/wav_tmp/WCE_output.txt > /dev/null 2>&1
 
 python /home/vagrant/repos/WCE_VM/aux_VM/WCE_to_rttm.py ${DATA_FOLDER}/wav_tmp/WCE_output.txt ${DATA_FOLDER}
+echo "WCE estimation complete."
